requirements.txt
fastapi==0.115.3
python-multipart==0.0.13
curl_cffi==0.7.3
uvicorn
tiktoken
python-dotenv
websockets
pillow
pybase64
jinja2
APScheduler
ua-generator
pyjwt
diskcache
sqlalchemy
alembic
passlib[bcrypt]
email-validator
